+++ test/all_files_test.dart
@@ -1,19 +1,2 @@
-import 'package:hyper_render/hyper_render.dart';
-import 'package:hyper_render/src/core/capture_extension.dart';
-import 'package:hyper_render/src/plugins/default_markdown_parser.dart';
-import 'package:hyper_render/src/plugins/default_code_highlighter.dart';
-import 'package:hyper_render/src/plugins/plugins.dart';
-import 'package:hyper_render/src/plugins/default_html_parser.dart';
-import 'package:hyper_render/src/plugins/default_css_parser.dart';
-import 'package:hyper_render/src/plugins/default_delta_parser.dart';
-import 'package:hyper_render/src/utils/svg_builder.dart';
-import 'package:hyper_render/src/utils/html_sanitizer.dart';
-import 'package:hyper_render/src/utils/html_heuristics.dart';
-import 'package:hyper_render/src/parser/delta/delta_adapter.dart';
-import 'package:hyper_render/src/parser/markdown/markdown_adapter.dart';
-import 'package:hyper_render/src/parser/html/html_adapter.dart';
-import 'package:hyper_render/src/parser/adapter.dart';
-import 'package:hyper_render/src/widgets/hyper_viewer.dart';
-import 'package:hyper_render/src/widgets/virtualized_selection_controller.dart';
-import 'package:hyper_render/src/widgets/virtualized_selection_overlay.dart';
+
 void main() {}